[Refact] AccountPage style 리팩토링
