+++ web/src/containers/AccountPage/styles/Background.js
@@ -1,14 +1,15 @@
 import styled from 'styled-components';
 
 const Background = styled.div`
+  /* Layout */
   display: flex;
   justify-content: center;
   align-items: center;
-
   position: absolute;
-  left: 0;
   top: 0;
+  left: 0;
 
+  /* Box */
   width: 100%;
   height: 100%;
 `;
