+++ web/src/containers/AccountPage/styles/HorizontalLine.js
@@ -1,6 +1,7 @@
 import styled from 'styled-components';
 
 const HorizontalLine = styled.hr`
+  /* Box */
   width: 90%;
   margin: 24px 0px;
 `;
