+++ web/src/containers/AccountPage/styles/SignHeader.js
@@ -1,10 +1,12 @@
 import styled from 'styled-components';
 
 const SignHeader = styled.header`
+  /* Layout */
   display: flex;
   flex-direction: column;
   align-items: center;
 
+  /* Box */
   width: 100%;
   margin: 10px 0px;
 `;
