+++ web/src/containers/AccountPage/styles/ValidationMessage.js
@@ -1,10 +1,13 @@
 import styled from 'styled-components';
 
 const ValidationMessage = styled.span`
-  color: red;
-  font-size: 0.7rem;
+  /* Box */
   width: 80%;
   margin-bottom: 4px;
+
+  /* Font */
+  font-size: 0.7rem;
+  color: red;
 `;
 
 export default ValidationMessage;
